Updated apis

+++ src/main/java/com/nitesh/appuploaderservice/repositories/AdjustmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdjustmentRepository extends JpaRepository<Adjustment, Long> {
+public interface AdjustmentRepository extends JpaRepository<Adjustment, Integer> {
 }
